+++ kadu-core/misc/date-time.h
@@ -1,7 +1,7 @@
 /*
  * %kadu copyright begin%
  * Copyright 2009, 2010, 2011 Piotr Galiszewski (piotr.galiszewski@kadu.im)
- * Copyright 2009, 2010 Rafał Malinowski (rafal.przemyslaw.malinowski@gmail.com)
+ * Copyright 2009, 2010, 2011 Rafał Malinowski (rafal.przemyslaw.malinowski@gmail.com)
  * %kadu copyright end%
  *
  * This program is free software; you can redistribute it and/or
@@ -29,6 +29,6 @@ class QDateTime;
 class QString;
 
 QString printDateTime(const QDateTime &datetime);
-void KADUAPI getTime(time_t *sec, int *msec);
+void KADUAPI getTime(time_t *sec, long int *msec);
 
 #endif // DATE_TIME
