+++ kadu-core/status/description-manager.h
@@ -1,8 +1,8 @@
 /*
  * %kadu copyright begin%
  * Copyright 2011 Piotr Galiszewski (piotr.galiszewski@kadu.im)
- * Copyright 2009, 2010, 2011 Rafał Malinowski (rafal.przemyslaw.malinowski@gmail.com)
  * Copyright 2012 Piotr Dąbrowski (ultr@ultr.pl)
+ * Copyright 2009, 2010, 2011 Rafał Malinowski (rafal.przemyslaw.malinowski@gmail.com)
  * %kadu copyright end%
  *
  * This program is free software; you can redistribute it and/or
@@ -34,7 +34,7 @@ class KADUAPI DescriptionManager : public QObject, public StorableStringList, pr
 
 	static DescriptionManager *Instance;
 
-	quint16 MaxNumberOfDescriptions;
+	int MaxNumberOfDescriptions;
 
 	DescriptionManager();
 	virtual ~DescriptionManager();
