+++ plugins/history/history-query.cpp
@@ -1,6 +1,9 @@
 /*
  * %kadu copyright begin%
- * Copyright 2012 Rafał Malinowski (rafal.przemyslaw.malinowski@gmail.com)
+ * Copyright 2009 Wojciech Treter (juzefwt@gmail.com)
+ * Copyright 2004 Adrian Smarzewski (adrian@kadu.net)
+ * Copyright 2007, 2008, 2009, 2010, 2011, 2012 Rafał Malinowski (rafal.przemyslaw.malinowski@gmail.com)
+ * Copyright 2004, 2006 Marcin Ślusarz (joi@kadu.net)
  * %kadu copyright end%
  *
  * This program is free software; you can redistribute it and/or
@@ -108,12 +111,12 @@ QDateTime HistoryQuery::toDateTime() const
 	return ToDateTime;
 }
 
-void HistoryQuery::setLimit(quint16 limit)
+void HistoryQuery::setLimit(int limit)
 {
 	Limit = limit;
 }
 
-quint16 HistoryQuery::limit() const
+int HistoryQuery::limit() const
 {
 	return Limit;
 }
