+++ plugins/history/history-query.h
@@ -1,6 +1,10 @@
 /*
  * %kadu copyright begin%
- * Copyright 2012 Rafał Malinowski (rafal.przemyslaw.malinowski@gmail.com)
+ * Copyright 2009 Wojciech Treter (juzefwt@gmail.com)
+ * Copyright 2004 Adrian Smarzewski (adrian@kadu.net)
+ * Copyright 2007, 2008, 2009, 2010, 2011, 2012 Rafał Malinowski (rafal.przemyslaw.malinowski@gmail.com)
+ * Copyright 2012 Bartosz Brachaczek (b.brachaczek@gmail.com)
+ * Copyright 2004, 2006 Marcin Ślusarz (joi@kadu.net)
  * %kadu copyright end%
  *
  * This program is free software; you can redistribute it and/or
@@ -51,7 +55,7 @@ class HISTORYAPI HistoryQuery
 	QDate ToDate;
 	QDateTime FromDateTime;
 	QDateTime ToDateTime;
-	quint16 Limit;
+	int Limit;
 
 public:
 	/**
@@ -177,14 +181,14 @@ public:
 	 *
 	 * Maximum limit items will be returned.
 	 */
-	void setLimit(quint16 limit);
+	void setLimit(int limit);
 
 	/**
 	 * @author Rafał 'Vogel' Malinowski
 	 * @short Return limit.
 	 * @return limit
 	 */
-	quint16 limit() const;
+	int limit() const;
 
 };
