+++ plugins/screenshot/gui/graphics-items/handler-rect-item.cpp
@@ -4,7 +4,8 @@
  * Copyright 2009 Wojciech Treter (juzefwt@gmail.com)
  * Copyright 2009 Bartłomiej Zimoń (uzi18@o2.pl)
  * Copyright 2004 Adrian Smarzewski (adrian@kadu.net)
- * Copyright 2007, 2008, 2009, 2010 Rafał Malinowski (rafal.przemyslaw.malinowski@gmail.com)
+ * Copyright 2007, 2008, 2009, 2010, 2011 Rafał Malinowski (rafal.przemyslaw.malinowski@gmail.com)
+ * Copyright 2013 Bartosz Brachaczek (b.brachaczek@gmail.com)
  * Copyright 2004, 2006 Marcin Ślusarz (joi@kadu.net)
  * %kadu copyright end%
  *
@@ -72,7 +73,7 @@ void HandlerRectItem::mouseMoveEvent(QGraphicsSceneMouseEvent *event)
 		return;
 	}
 
-	emit movedTo(Type, event->scenePos().x(), event->scenePos().y());
+	emit movedTo(Type, static_cast<int>(event->scenePos().x()), static_cast<int>(event->scenePos().y()));
 }
 
 QRectF HandlerRectItem::boundingRect() const
@@ -90,3 +91,5 @@ void HandlerRectItem::paint(QPainter *painter, const QStyleOptionGraphicsItem *o
 	painter->setPen(Qt::black);
 	painter->drawRect(QRect(0, 0, Size, Size));
 }
+
+#include "moc_handler-rect-item.cpp"
