+++ plugins/screenshot/notify/screenshot-notification.cpp
@@ -2,8 +2,8 @@
  * %kadu copyright begin%
  * Copyright 2010, 2011 Piotr Galiszewski (piotr.galiszewski@kadu.im)
  * Copyright 2011 Piotr Dąbrowski (ultr@ultr.pl)
- * Copyright 2010 Rafał Malinowski (rafal.przemyslaw.malinowski@gmail.com)
- * Copyright 2010, 2011 Bartosz Brachaczek (b.brachaczek@gmail.com)
+ * Copyright 2010, 2011 Rafał Malinowski (rafal.przemyslaw.malinowski@gmail.com)
+ * Copyright 2010, 2011, 2013 Bartosz Brachaczek (b.brachaczek@gmail.com)
  * %kadu copyright end%
  *
  * This program is free software; you can redistribute it and/or
@@ -46,7 +46,7 @@ void ScreenshotNotification::unregisterNotifiactions()
 	SizeLimitNotification = 0;
 }
 
-void ScreenshotNotification::notifySizeLimit(int size)
+void ScreenshotNotification::notifySizeLimit(long size)
 {
 	ScreenshotNotification *notification = new ScreenshotNotification();
 	notification->setTitle(tr("ScreenShot size limit"));
@@ -63,3 +63,5 @@ ScreenshotNotification::~ScreenshotNotification()
 {
 
 }
+
+#include "moc_screenshot-notification.cpp"
