+++ plugins/screenshot/notify/screenshot-notification.h
@@ -1,10 +1,11 @@
 /*
  * %kadu copyright begin%
  * Copyright 2008, 2011 Piotr Galiszewski (piotr.galiszewski@kadu.im)
+ * Copyright 2012 Wojciech Treter (juzefwt@gmail.com)
  * Copyright 2008 Michał Podsiadlik (michal@kadu.net)
  * Copyright 2009 Bartłomiej Zimoń (uzi18@o2.pl)
  * Copyright 2004 Adrian Smarzewski (adrian@kadu.net)
- * Copyright 2007, 2008, 2009, 2010 Rafał Malinowski (rafal.przemyslaw.malinowski@gmail.com)
+ * Copyright 2007, 2008, 2009, 2010, 2011 Rafał Malinowski (rafal.przemyslaw.malinowski@gmail.com)
  * Copyright 2004, 2006 Marcin Ślusarz (joi@kadu.net)
  * %kadu copyright end%
  *
@@ -25,7 +26,7 @@
 #ifndef SCREENSHOT_NOTIFICATION_H
 #define SCREENSHOT_NOTIFICATION_H
 
-#include "notify/notification.h"
+#include "notify/notification/notification.h"
 
 class NotifyEvent;
 
@@ -39,7 +40,7 @@ public:
 	static void registerNotifications();
 	static void unregisterNotifiactions();
 
-	static void notifySizeLimit(int size);
+	static void notifySizeLimit(long size);
 
 	explicit ScreenshotNotification();
 	virtual ~ScreenshotNotification();
